dashboard/backend/api/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
soundfile
scipy
numpy
Pillow
